KOGITO-1609: Moving Tests
